No action hard fork necessary
- Move `ClaimStakeReward4` to `ClaimStakeReward3`.
- Only internal structure and stake data condition are changed. Reward
  logic itself is not changed.
- Move TableDataVersionConfig to StakeState.
