--- Lib9c/TableData/TableDataVersionConfig.cs
@@ -1,9 +1,0 @@
-namespace Nekoyume.TableData
-{
-    public static class TableDataVersionConfig
-    {
-        // PatchTableSheet version update without changing model version
-        // FIXME: Set block index to real block index to change reward version
-        public const long StakeRewardSheetV2Index = 7_000_000L;
-    }
-}
